Combined patches - Make GrowthBookSDK gb property optional
This was a late property before which was causing issues because the
UI would load before gb was initialized. This caused the screen to
flash red with a LateInitializationError. Once gb was initialized the
red screen would be replaced with the example app main screen.

Check gb != null before loading UI.

Change properties from being late initialized to empty maps

When the device has no internet connection we are still trying to use
these properties, but they have not been initialized. We can fix this
by initializing them as empty maps to prevent a LateInitializationError.
Then we check .isNotEmpty when setting the model.

Add “features” to GBSDKBuilderApp constructor

Add test that makes sure we get features even when the network
connection fails or we get a bad response.
Change late properties to be optional. When they are late and we get
a failed network response it results in a LateInitializationError and
freezes the app.
If model is null, return out of fetchFeature so we don't reset the
seeded features we added before the request. This way we can still rely
on them within the project logic.

<!-- ps-id: ba55858b-c24d-4a9f-b864-0d8cff476d4c -->