Fix LateInitialization in FeatureDataSource
Prior to this patch when FeatureDataSource would fail due to a client
failure it would incorrectly result in a LateInitialization exception.
This was because FeatureDataSource was written assuming the client was
Future<> based but the client was actually Callback based. Having these
two different async approaches is not beneficial and is what enables
this LateInitialization issue to even be possible.

To resolve this issue I changed the abstract BaseClient interface to use
the newer Future<> style async rather than the older style Callback
approach.

I then proceeded to update the FeatureDataSource to properly use the new
Future<> based client properly and also modified it to eliminate it's
needless propagation of the onError callback. I then went through and
made sure that all the uses of FeatureDataSource were properly updated.
I then made sure to even wrap the usage of FeatureDataSource is in a
swallowing try/catch to maintain the existing external behavior. We made
decide to change this. But figured it best to not change external
behavior of the API as that would be a breaking change.

In my opinion changing the behavior to NOT swallow errors would be a
better long term approach IMHO. It would enable users to, for one, be
aware of the errors that are happening. And two, be able to explicitly
handle or ignore the errors. But, to expedite getting this change in I
will leave this for a future patch where more time and discussion can be
had around it. As it would be a breaking change.

[changelog]
fixed: LateInitialization exception on client failures in FeatureDataSource

<!-- ps-id: 03e656a4-84d9-46ce-8622-064fb0652cf9 -->